requirements.txt
Django==4.0.2
django-ckeditor==6.2.0
html2text==2020.1.16
python-telegram-bot==13.11